+++ packages/components/organisms/f-status-banner/test-utils/component-objects/f-status-banner.selectors.js
@@ -1,3 +1,9 @@
-export const ADDRESS_BOX = '[data-test-id="address-box-input"]';
-export const BUTTON = '[data-test-id="find-restaurants-button"]';
-export const ERROR_MESSAGE = '[id="errorMessage"]';
+const ADDRESS_BOX = '[data-test-id="address-box-input"]';
+const BUTTON = '[data-test-id="find-restaurants-button"]';
+const ERROR_MESSAGE = '[id="errorMessage"]';
+
+export {
+    ADDRESS_BOX,
+    BUTTON,
+    ERROR_MESSAGE
+};