+++ packages/storybook/CHANGELOG.md
@@ -4,6 +4,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.52.29
+------------------------------
+*July 07, 2022*
+### Changed
+- updated vue.config to include f-contact-preferences in components using @use and @forward
+
+
+v0.52.28
+------------------------------
+*July 07, 2022*
+### Changed
+- updated vue.config to include f-checkout in components using @use and @forward
+
+
 v0.52.27
 ------------------------------
 *July 06, 2022*